code

+++ app/services/addresses_service.rb
@@ -1,5 +1,9 @@
 class AddressesService < BaseService
-  def get_addresses(customer)
+  def get_addresses(query_params)
+    if (customer = Customer.find_by(id: query_params["customer_id"])).blank?
+      return CommonService.response_format(ResponseCode.COMMON.FAILED,
+                                           "ERROR: customer:#{customer.inspect} is blank!")
+    end
     CommonService.response_format(ResponseCode.COMMON.OK, customer.addresses)
   end
 
@@ -7,8 +11,20 @@ class AddressesService < BaseService
     CommonService.response_format(ResponseCode.COMMON.OK, address)
   end
 
-  def create_address(customer, address_params)
+  def create_address(address_params)
+    # 参数合法性检查
+    if address_params.blank?
+      return CommonService.response_format(ResponseCode.COMMON.FAILED,
+                                           "ERROR: address_params:#{address_params.inspect} is blank!")
+    end
+
     Address.transaction do
+      customer_id = address_params.extract!("customer_id")["customer_id"]
+      if (customer = Customer.find_by(id: customer_id)).blank?
+        return CommonService.response_format(ResponseCode.COMMON.FAILED,
+                                             "ERROR: customer:#{customer.inspect} is blank!")
+      end
+
       # 处理默认地址唯一性
       if address_params[:is_default] == true
         customer.addresses.collect{|address| address.update(is_default: false)}
@@ -20,7 +36,16 @@ class AddressesService < BaseService
   end
 
   def update_address(address, address_params)
+    # 参数合法性检查
+    if address.blank? || address_params.blank?
+      return CommonService.response_format(ResponseCode.COMMON.FAILED,
+                                           "ERROR: address_params:#{address_params.inspect} is blank!")
+    end
+
     Address.transaction do
+      # 不能修改所属用户，过滤掉customer外键。
+      address_params.extract!("customer_id")
+
       # 处理默认地址唯一性
       if address_params[:is_default] == true
         Address.where(customer_id: address.customer_id).collect{|address| address.update(is_default: false)}
@@ -36,8 +61,9 @@ class AddressesService < BaseService
   end
 
   def destory_address(address)
-    Address.where(customer_id: address.customer_id).first.update(is_default: true)
-    address.destroy
+    if !address.nil?
+      address.destroy
+    end
     CommonService.response_format(ResponseCode.COMMON.OK)
   end
